requirements.txt
flask
tflite-runtime
Pillow
flask-cors
numpy